voice bot/requirements.txt
fastapi
uvicorn
websockets
google-generativeai
python-dotenv
gTTS
SpeechRecognition
pydub
